Simplify API error types
Return `Future<String>` rather than `Future<Either<String, String>>`, by
returning a failed future if the API update fails.

This preserves the full stack trace if the API request throws an exception,
and makes it easier to handle the errors in the collect method because
there is only one type of error: a failed Future.