Use updated input names.
I changed the input type names in the API. This worked when making
requests through postman but I think sangria is doing some proper type
checking so this update is needed to get it to work with the API again.

This will fix integration. We'll have to be careful deploying this
change and the API on prod because it will break if they're not deployed
together.